保存用户操作日志

+++ src/main/java/com/anthonyzero/scaffold/system/entity/SysLog.java
@@ -47,7 +47,7 @@ public class SysLog implements Serializable {
      * 耗时
      */
     @TableField("time")
-    private BigDecimal time;
+    private Long time;
 
     /**
      * 方法
